+++ tests/TestCase/Utility/LinkScannerTest.php
@@ -144,8 +144,6 @@ class LinkScannerTest extends TestCase
             if ($response->isOk()) {
                 $this->assertNotEmpty($responseFromCache);
                 $this->assertInstanceof(Response::class, $responseFromCache);
-            } else {
-                $this->assertEmpty($responseFromCache);
             }
         }
 
@@ -174,8 +172,8 @@ class LinkScannerTest extends TestCase
             null => $this->LinkScanner->getConfig('target') . DS . 'results_' . $this->LinkScanner->hostname . '_' . $this->LinkScanner->startTime,
             'example' => $this->LinkScanner->getConfig('target') . DS . 'example',
             TMP . 'example' => TMP . 'example',
-        ] as $filenameWhereToExport => $expectedFilename) {
-            $result = $this->LinkScanner->export($filenameWhereToExport);
+        ] as $filename => $expectedFilename) {
+            $result = $this->LinkScanner->export($filename);
             $this->assertFileExists($result);
             $this->assertEquals($expectedFilename, $result);
             $this->assertEventFired('LinkScanner.resultsExported', $this->EventManager);
@@ -221,13 +219,13 @@ class LinkScannerTest extends TestCase
 
         //Gets properties from both client, fixes properties of the `Client`
         //  instances and performs the comparison
-        $expectedClientProperties = $this->getProperties($this->LinkScanner->Client);
-        $resultClientProperties = $this->getProperties($result->Client);
+        $expectedProperties = $this->getProperties($this->LinkScanner->Client);
+        $resultProperties = $this->getProperties($result->Client);
         foreach (['_adapter', '__phpunit_invocationMocker', '__phpunit_originalObject', '__phpunit_configurable'] as $key) {
-            unset($expectedClientProperties[$key], $resultClientProperties[$key]);
+            unset($expectedProperties[$key], $resultProperties[$key]);
         }
-        $expectedProperties = ['Client' => $expectedClientProperties] + $this->getProperties($this->LinkScanner);
-        $resultProperties = ['Client' => $resultClientProperties] + $this->getProperties($result);
+        $expectedProperties = ['Client' => $expectedProperties] + $this->getProperties($this->LinkScanner);
+        $resultProperties = ['Client' => $resultProperties] + $this->getProperties($result);
         $this->assertEquals($expectedProperties, $resultProperties);
 
         //With a no existing file
@@ -302,12 +300,7 @@ class LinkScannerTest extends TestCase
         $hostname = get_hostname_from_url($this->fullBaseUrl);
 
         foreach ($LinkScanner->ResultScan as $item) {
-            if (!$item->external) {
-                $this->assertRegexp(sprintf('/^https?:\/\/%s/', preg_quote($hostname)), $item->url);
-            } else {
-                $this->assertTextStartsNotWith('http://' . $hostname, $item->url);
-                $this->assertTextStartsNotWith('https://' . $hostname, $item->url);
-            }
+            $this->assertRegexp(sprintf('/^https?:\/\/%s/', preg_quote($hostname)), $item->url);
             $this->assertContains($item->code, [200, 500]);
             $this->assertStringStartsWith('text/html', $item->type);
         }
@@ -373,7 +366,7 @@ class LinkScannerTest extends TestCase
         $this->assertEquals($expectedDebug, $this->debug);
 
         //Results contain both internal and external urls
-        $expectedInternaLinks = [
+        $expectedInternal = [
             'http://localhost',
             'http://localhost/pages/first_page',
             'http://localhost/favicon.ico',
@@ -384,11 +377,11 @@ class LinkScannerTest extends TestCase
             'http://localhost/pages/redirect',
             'http://localhost/pages/sameredirect',
         ];
-        $expectedExternalLinks = ['http://google.it'];
+        $expectedExternal = ['http://google.it'];
         $internalLinks = $LinkScanner->ResultScan->match(['external' => false])->extract('url');
         $externalLinks = $LinkScanner->ResultScan->match(['external' => true])->extract('url');
-        $this->assertEquals($expectedInternaLinks, $internalLinks->toList());
-        $this->assertEquals($expectedExternalLinks, $externalLinks->toList());
+        $this->assertEquals($expectedInternal, $internalLinks->toList());
+        $this->assertEquals($expectedExternal, $externalLinks->toList());
 
         $this->debug = [];
 
@@ -419,13 +412,13 @@ class LinkScannerTest extends TestCase
         $LinkScanner->setConfig('followRedirects', true)->scan();
         $this->assertEquals($expectedDebug, $this->debug);
 
-        array_pop($expectedInternaLinks);
-        array_pop($expectedInternaLinks);
-        $expectedInternaLinks[] = 'http://localhost/pages/third_page';
+        array_pop($expectedInternal);
+        array_pop($expectedInternal);
+        $expectedInternal[] = 'http://localhost/pages/third_page';
         $internalLinks = $LinkScanner->ResultScan->match(['external' => false])->extract('url');
         $externalLinks = $LinkScanner->ResultScan->match(['external' => true])->extract('url');
-        $this->assertEquals($expectedInternaLinks, $internalLinks->toList());
-        $this->assertEquals($expectedExternalLinks, $externalLinks->toList());
+        $this->assertEquals($expectedInternal, $internalLinks->toList());
+        $this->assertEquals($expectedExternal, $externalLinks->toList());
 
         $LinkScanner = $this->getLinkScannerClientReturnsFromTests($params);
         $LinkScanner->setConfig('maxDepth', 1)->scan();
@@ -459,6 +452,6 @@ class LinkScannerTest extends TestCase
         $EventManager = $this->getEventManager($LinkScanner);
         $LinkScanner->scan();
 
-        $this->assertEventFired('LinkScanner.' . 'responseNotOk', $EventManager);
+        $this->assertEventFired('LinkScanner.responseNotOk', $EventManager);
     }
 }